+++ src/lavalink/mod.rs
@@ -1,6 +1,6 @@
 //! # Hydrolink
 //!
-//! An [tokio](https://tokio.rs) based [Lavalink](https://lavalink.dev/) client, with support for any Discord library.
+//! A [tokio](https://tokio.rs) based [Lavalink](https://lavalink.dev/) client, with support for any Discord library.
 
 pub mod cluster;
 pub mod hydrogen;
@@ -14,7 +14,6 @@ pub use model::*;
 pub use rest::*;
 use tokio::net::TcpStream;
 use tokio_tungstenite::{MaybeTlsStream, WebSocketStream};
-pub use websocket::*;
 
 /// User agent for the REST client.
 pub const LAVALINK_USER_AGENT: &str =
@@ -34,36 +33,68 @@ pub type LavalinkConnection = WebSocketStream<MaybeTlsStream<TcpStream>>;
 pub enum Error {
     /// An error from [`reqwest`].
     Reqwest(reqwest::Error),
+
     /// An error from [`serde_json`].
     Serde(serde_json::Error),
+
+    #[cfg(feature = "simd-json")]
+    /// An error from [`simd_json`].
+    SimdJson(simd_json::Error),
+
     /// An error from the Lavalink server.
     Lavalink(model::Error),
+
     /// An error from [`http`].
     Http(http::Error),
+
     /// An error from [`tokio_tungstenite`].
     Tungstenite(tokio_tungstenite::tungstenite::Error),
+
+    UrlParse(url::ParseError),
+
     /// No session ID was provided.
     NoSessionId,
+
     /// The message received from the Lavalink server was invalid.
     InvalidMessage,
+
     /// The password provided to the Lavalink server was invalid.
     InvalidHeaderValue(InvalidHeaderValue),
+
     /// The Lavalink node is already connected.
     AlreadyConnected,
+
+    /// The response from the Lavalink server doesn't have a body.
+    NoResponseBody,
 }
 
 impl std::fmt::Display for Error {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Self::Reqwest(e) => write!(f, "Reqwest error: {}", e),
-            Self::Serde(e) => write!(f, "Serde error: {}", e),
-            Self::Lavalink(e) => write!(f, "Lavalink error: {:?}", e),
-            Self::Http(e) => write!(f, "HTTP error: {}", e),
-            Self::Tungstenite(e) => write!(f, "Tungstenite error: {}", e),
-            Self::InvalidHeaderValue(e) => write!(f, "Invalid header value: {}", e),
+            Self::Reqwest(e) => e.fmt(f),
+
+            Self::Serde(e) => e.fmt(f),
+
+            Self::Lavalink(e) => write!(f, "Lavalink REST error: {}", e.message),
+
+            Self::Http(e) => e.fmt(f),
+
+            Self::Tungstenite(e) => e.fmt(f),
+
+            Self::InvalidHeaderValue(e) => e.fmt(f),
+
+            Self::UrlParse(e) => e.fmt(f),
+
             Self::NoSessionId => write!(f, "No session ID was provided"),
+
             Self::InvalidMessage => write!(f, "Lavalink sent an invalid message"),
+
             Self::AlreadyConnected => write!(f, "Lavalink node is already connected"),
+
+            Self::NoResponseBody => write!(f, "Lavalink response had no body"),
+
+            #[cfg(feature = "simd-json")]
+            Self::SimdJson(e) => e.fmt(f),
         }
     }
 }
@@ -98,9 +129,22 @@ impl From<tokio_tungstenite::tungstenite::Error> for Error {
     }
 }
 
+impl From<url::ParseError> for Error {
+    fn from(e: url::ParseError) -> Self {
+        Self::UrlParse(e)
+    }
+}
+
 impl From<InvalidHeaderValue> for Error {
     fn from(e: InvalidHeaderValue) -> Self {
         Self::InvalidHeaderValue(e)
+    }
+}
+
+#[cfg(feature = "simd-json")]
+impl From<simd_json::Error> for Error {
+    fn from(e: simd_json::Error) -> Self {
+        Self::SimdJson(e)
     }
 }